requirements.txt
dagster
psycopg2-binary
requests
python-dotenv

pandas
